"check_style_score -> 1 of 3 passed. style score: 0/1 "

+++ cssp2-Assignments/m23/Assignment-1/Distance.java
@@ -94,7 +94,7 @@ public class Distance {
 		}
 		return innerproductvalue;
 	}
-	double angle(TreeMap<String, Integer> comparedict) {
+	long angle(TreeMap<String, Integer> comparedict) {
 		double compnorm = 0;
 		for (int i : comparedict.values()) {
 			compnorm = compnorm + (i * i);
